Alteração da tipagem da variável de short para int para permitir gerar mais mensagens via POST

+++ kafka-producer/Models/Pedido.cs
@@ -2,6 +2,6 @@ namespace kafka_producer.Models
 {
     public class Pedido
     {
-        public short Mensagens { get; set; }
+        public int Mensagens { get; set; }
     }
 }
